requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
kafka-python==1.4.7
MarkupSafe==1.1.1
numpy==1.18.0
opencv-contrib-python==4.1.2.30
Werkzeug==0.16.0
